// src/app/Models/ICompany.ts
    export interface ICompany {
        id: number;
        name: string;
        tradingName: string;
        industry: string;
        organisationType: string;
        gst: string;
        website: string;
        timezone: string;
        logo: string;
        address: string;
        suburb: string;
        city: string;
        state: string;
        country: string;
        postCode: string;
        firstName: string;
        lastName: string;
        officePhone: string;
        mobileNumber: string;
        email: string;
        note: string;
        validDate: Date;
        initialized: boolean;
    }